Refactor: database, libs 의 common으로 이동
