+++ libs/common/src/database/seed.ts
@@ -1,8 +1,8 @@
 import { NestFactory } from '@nestjs/core';
 import { Logger } from '@nestjs/common';
 
-import { SeederModule } from '@api/database/seeder/seeder.module';
-import { Seeder } from '@api/database/seeder/seeder';
+import { SeederModule } from '@app/common/database/seeder/seeder.module';
+import { Seeder } from '@app/common/database/seeder/seeder';
 
 async function bootstrap() {
   NestFactory.createApplicationContext(SeederModule)
