+++ libs/common/src/database/seeder/seeder.module.ts
@@ -1,12 +1,12 @@
 import { Module, Logger } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 
-import { DatabaseModule } from '@api/database/database.module';
+import { DatabaseModule } from '@app/common/database/database.module';
 import { Seeder } from './seeder';
 import { UserSeederModule } from './user/user-seeder.module';
 import { CategorySeederModule } from './category/category-seeder.module';
 import { ArticleSeederModule } from './article/article-seeder.module';
-import { ormconfig } from '@api/database/ormconfig';
+import { ormconfig } from '@app/common/database/ormconfig';
 
 @Module({
   imports: [
